Added the Sorting Capability

+++ src/ca/ottawaparking/Bike.java
@@ -120,7 +120,7 @@ public class Bike{
 		return null;
 	}
 	
-	public float getDist(float latin, float longin){
+	public float getDist(double latin, double longin){
 		// Calculates the distance between the current bike stop and another coordinate
 		return (float)(Math.sqrt(Math.pow((longin-lon),2) + Math.pow((latin - lat), 2)));
 	}
